requirements.txt
# Core
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3

# ML Models
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.2

# Data Collection
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0
webdriver-manager==4.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # Per PostgreSQL (opzionale)

# Utils
python-dotenv==1.0.0
tqdm==4.65.0
loguru==0.7.2
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.5.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Notebooks
jupyter==1.0.0
ipykernel==6.25.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Code Quality
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0